+++ app/Http/Controllers/commentairesController.php
@@ -54,8 +54,6 @@ class commentairesController extends Controller
 
                     $datas['date_commentaire'] = date_create(now());
 
-                    // $datas['date_commentaire'] =  date('Format String', time());
-
                     $commentaires = commentaires::create($datas);
                     
                     if ($commentaires) {
@@ -91,18 +89,21 @@ class commentairesController extends Controller
     public function Commentaires(){
 
         $commentaires = commentaires::join('utilisateurs', 'utilisateurs.id', '=', 'commentaires.utilisateurs_id')
-            ->join('annonces', 'annonces.id', '=', 'commentaires.annonces_id')
-            ->join('sous_categories', 'annonces.sous_categories_id', '=', 'sous_categories.id')
-            ->join('categories', function($join)
-                {
-                    $join->on('categories.id', '=', 'sous_categories.categories_id');
-                })
-            ->select('commentaires.id',
-                'commentaires.commentaire',
-                'commentaires.date_commentaire',
-                'annonces.titre',
-                'utilisateurs.login',
-            )->get();
+        ->join('annonces', function($join)
+            {
+                $join->on('annonces.id', '=', 'commentaires.annonces_id');
+            })
+        ->join('sous_categories', 'annonces.sous_categories_id', '=', 'sous_categories.id')
+        ->join('categories', function($join)
+            {
+                $join->on('categories.id', '=', 'sous_categories.categories_id');
+            })
+        ->select('commentaires.id',
+            'commentaires.commentaire',
+            'commentaires.date_commentaire',
+            'annonces.titre',
+            'utilisateurs.login',
+        )->get();
 
         if ($commentaires) {
 
@@ -133,33 +134,41 @@ class commentairesController extends Controller
 
         $commentaires = commentaires::where('commentaires.id', '=', $id)
         ->join('utilisateurs', 'utilisateurs.id', '=', 'commentaires.utilisateurs_id')
-        ->select(
-            'commentaires.id',
+        ->join('annonces', function($join)
+            {
+                $join->on('annonces.id', '=', 'commentaires.annonces_id');
+                // ->where('annonces.actif', '=', true)
+                // ->where('annonces.etat', '=', true);
+            })
+        ->join('sous_categories', 'annonces.sous_categories_id', '=', 'sous_categories.id')
+        ->join('categories', function($join)
+            {
+                $join->on('categories.id', '=', 'sous_categories.categories_id');
+            })
+        ->select('commentaires.id',
             'commentaires.commentaire',
             'commentaires.date_commentaire',
+            'annonces.titre',
             'utilisateurs.login',
         )->get();
 
         if ($commentaires) {
-            
+
             return response([
-                
                 'code' => '200',
                 'message' => 'success',
-                'data' => $commentaires,
-                
+                'data' => $commentaires
             ], 200);
 
         }else {
-            
+
             return response([
-                'code' => '004',
-                'message' => 'Identifiant incorrect',
-                'data' => null
+                'code' => '005',
+                'message' => 'La table est vide',
+                'data' => $commentaires
             ], 201);
 
         }
-
     }
 
 
@@ -167,12 +176,29 @@ class commentairesController extends Controller
 
     public function Utilisateur($id){
 
-        $utilisateur = commentaires::find($id)->Utilisateurs;
+        // $utilisateur = commentaires::find($id)->Utilisateurs;
+
+        $utilisateur = commentaires::where('commentaires.id', '=', $id)
+        ->join('utilisateurs', function($join)
+            {
+                $join->on('utilisateurs.id', '=', 'commentaires.utilisateurs_id')
+                ->where('utilisateurs.actif', '=', true);
+            })
+        ->select('utilisateurs.id',
+            'utilisateurs.login',
+            'utilisateurs.email',
+            'utilisateurs.photo',
+            'utilisateurs.role',
+            'utilisateurs.date_creation',
+            'utilisateurs.nomAdministrateur',
+            'utilisateurs.prenomAdministrateur',
+            'utilisateurs.telephoneAdministrateur',
+        )->get();
 
         if ($utilisateur) {
             
             return response([
-                'code' => 'success',
+                'code' => '200',
                 'message' => 'success',
                 'data' => $utilisateur
             ], 200);
@@ -214,8 +240,6 @@ class commentairesController extends Controller
                     'annonces.sous_categories_id',
                     'sous_categories.nom_sous_categorie',
                     'categories.nomCategorie',
-                    'categories.image',
-                    'categories.titre'
                 )->get();
     
             if ($commentaires) {
@@ -243,45 +267,63 @@ class commentairesController extends Controller
 
     public function putCommentaire(Request $request, $id){
 
-        $identif = commentaires::findOrFail($id);
-
-        $data = $request->all();
-
-        $validator = Validator::make($request->all(), [
+        if (Auth::check()) {
             
-            'commentaire' => 'required',
-            'annonces_id' => 'required'
-        ]);
+            $user = Auth::user();
 
-        if ($validator->fails()) {
+            $idUser = Auth::id();
 
-            $erreur = $validator->errors();
-            
-            return response([
-                'code' => '001',
-                'message' => 'L\'un des champs est vide ou ne respecte pas le format',
-                'data' => $erreur
-            ], 201);
+            $role = $user['role'];
 
-        }else {
+            $donnees = commentaires::where('utilisateurs_id', '=', $idAuth)->addSelect('id')->first();
 
-            $commentaires = $identif->update($data);
+            $idU = $donnees['id'];
 
-            if ($commentaires) {
-                
-                return response([
-                    'code' => '200',
-                    'message' => 'success',
-                    'data' => $identif
-                ], 200);
+            if ( ($idUser == $idU) || $role == "administrateur") {
 
-            }else {
-                
-                return response([
-                    'code' => '005',
-                    'message' => 'Echec lors de l\'opération',
-                    'data' => null
-                ], 201);
+                $identif = commentaires::findOrFail($id);
+
+                $data = $request->all();
+
+                $validator = Validator::make($request->all(), [
+                    
+                    'commentaire' => 'required',
+                    'annonces_id' => 'required'
+                ]);
+
+                if ($validator->fails()) {
+
+                    $erreur = $validator->errors();
+                    
+                    return response([
+                        'code' => '001',
+                        'message' => 'L\'un des champs est vide ou ne respecte pas le format',
+                        'data' => $erreur
+                    ], 201);
+
+                }else {
+
+                    $commentaires = $identif->update($data);
+
+                    if ($commentaires) {
+                        
+                        return response([
+                            'code' => '200',
+                            'message' => 'success',
+                            'data' => $identif
+                        ], 200);
+
+                    }else {
+                        
+                        return response([
+                            'code' => '005',
+                            'message' => 'Echec lors de l\'opération',
+                            'data' => null
+                        ], 201);
+
+                    }
+                    
+                }
 
             }
             
@@ -290,31 +332,57 @@ class commentairesController extends Controller
     }
 
 
+    
     // Supprimer une commentaire
      
     public function deleteCommentaire($id){
 
-        $delete = commentaires::findOrFail($id)->delete();
+        if (Auth::check()) {
+            
+            $user = Auth::user();
 
-        if ($delete) {
+            $idUser = Auth::id();
 
-            return response([
-                'code' => '200',
-                'message' => 'Suppression effectuée avec succes',
-                'data' => null
-            ], 200);
+            $role = $user['role'];
 
-        } else {
+            $donnees = commentaires::where('utilisateurs_id', '=', $idAuth)->addSelect('id')->first();
 
-            return response([
-                'code' => '004',
-                'message' => 'L\'identifiant incorrect',
-                'data' => null
-            ], 201);
+            $idU = $donnees['id'];
 
+            if ( ($idUser == $idU) || $role == "administrateur") {
+
+                $delete = commentaires::findOrFail($id)->delete();
+
+                if ($delete) {
+
+                    return response([
+                        'code' => '200',
+                        'message' => 'Suppression effectuée avec succes',
+                        'data' => null
+                    ], 200);
+
+                } else {
+
+                    return response([
+                        'code' => '004',
+                        'message' => 'L\'identifiant incorrect',
+                        'data' => null
+                    ], 201);
+
+                }
+
+            }else{
+
+                return response([
+                    'code' => '004',
+                    'message' => 'Acces non autorise',
+                    'data' => null
+                ], 201);
+
+            }
+            
         }
         
-    }
-    
+    } 
 
 }
